% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstStage.R
\name{firstStage}
\alias{firstStage}
\title{First stage of the two-stage method}
\usage{
firstStage(ylow, n, p, seas_levels, regMat = NULL, plotFlag = TRUE)
}
\arguments{
\item{ylow}{averaged data.}

\item{n}{Number of samples in the model period.}

\item{p}{Number of samples in the holdout period.}

\item{seas_levels}{periods corresponding to levels of
seasonality in the data}

\item{regMat}{covariates to be used for regression.}

\item{plotFlag}{plot results. default is TRUE.}
}
\description{
Fits time series models for the averaged (Y_low)
    data. It computes and returns five possible models and
    their preidctions.
}
\examples{

}
