% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondStage.R
\name{secondStage}
\alias{secondStage}
\title{Seasonality estimation for high frequency time series}
\usage{
secondStage(d, seas_periods)
}
\arguments{
\item{d}{the actual data.}

\item{seas_periods}{periods associated with the levels of seasonality}
}
\description{
Estimates the seasonality using different methods like
classical decomposition method based, polynomial fitting,
trigonometric methods and spline.
}
\examples{

}
