% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{twoStage.accuracy}
\alias{twoStage.accuracy}
\title{Find the accuracy of a forecast}
\usage{
twoStage.accuracy(y, yhat)
}
\arguments{
\item{y}{the actual data.}

\item{yhat}{the forecasted vector.}
}
\description{
Calculates the Mean Square Error (MSE), Mean Absolute
    Deviation (MAD) and Mean Absolute Percentage Error (MAPE) of
    a forecast.
}
\examples{

}
