% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical.R
\name{twoStage.classical}
\alias{twoStage.classical}
\title{Classical decomposition method for seasonality estimation}
\usage{
twoStage.classical(data)
}
\arguments{
\item{y}{the actual data.}

\item{yhat}{the forecasted vector.}
}
\description{
Estimates the seasonality at different levels by
    subtracting the mean for additive seasonality
    (dividing for multiplicative) and then averaging across
    the entire data.
}
\examples{

}
