% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalizedLikelihood.R
\name{twoStage.penalizedLikelihood}
\alias{twoStage.penalizedLikelihood}
\title{Metric for model selection}
\usage{
twoStage.penalizedLikelihood(MSE, n, p)
}
\arguments{
\item{MSE}{Mean Square Error of the fitted model.}

\item{n}{length of the data.}

\item{p}{number of parameters estimated by the model.}
}
\description{
Finds psi = n x ln(MSE) + 2 x p, an alternative to AIC.
    psi, or penalized lack-of-fit, is used for model selection.
}
\examples{

}
