% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{twoStage.spline}
\alias{twoStage.spline}
\title{splines for seasonality estimation}
\usage{
twoStage.spline(x)
}
\arguments{
\item{x}{the actual data.}
}
\description{
Estimates the seasonality for an input function
    by using splines
}
\examples{

}
