% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.R
\name{twoStage.trig}
\alias{twoStage.trig}
\title{Trigonometric estimation of seasonality}
\usage{
twoStage.trig(x)
}
\arguments{
\item{x}{data whose seasonality is to be estimated.}
}
\description{
Estimate the seasonality using trigonometric functions.
}
\examples{

}
