% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical.R
\name{twoStage.classical}
\alias{twoStage.classical}
\title{Classical decomposition method for seasonality estimation}
\usage{
twoStage.classical(d, period)
}
\arguments{
\item{d}{the actual data.}

\item{period}{seasonality levels.}
}
\description{
Estimates the seasonality at different levels by
subtracting the mean for additive seasonality
(dividing for multiplicative) and then averaging across
the entire data.
}
\examples{

}
